+++ Functions/Assertions/Be.Tests.ps1
@@ -7,6 +7,9 @@ Describe "PesterBe" {
     It "returns true if the 2 arguments are equal" {
         Test-PositiveAssertion (PesterBe 1 1)
     }
+	It "returns true if the 2 arguments are equal and have different case" {
+        Test-PositiveAssertion (PesterBe "A" "a")
+    }
 
     It "returns false if the 2 arguments are not equal" {
         Test-NegativeAssertion (PesterBe 1 2)