+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   - Fixed mix scoping of It and Context [GH-98] and [GH-99]
   - Test Drives are randomly generated, which should allow concurrent Pester processes [GH-100] and [GH-94] 
   - Fixed nUnit test failing on non-US computers [GH-109]
+  - Add case sensitive Be, Contain and Match assertions [GH-107]
 
 ## 2.0.3 (Apr 16, 2013)